split files to modules, create makefile.
